myproject/requirements.txt
Django
djangorestframework
django-cors-headers
PyJWT
selenium
webdriver-manager
